+++ src/Superstars.DB/Scripts/20180529.006.Install-sBankRollUpdate.sql
@@ -1,5 +1,5 @@
 create proc sp.sBankRollUpdate(
-	@RealCoins decimal(15,8),
+	@RealCoins int,
 	@FakeCoins int
 ) as
 begin
